feat: screenshot handler
